requirements.txt
pandas
langchain
langchain-groq
langchain-openai
streamlit
python-dotenv
pydantic